fix: Link preview and search should work on collection descriptions
